Changes in how templates are applied.
It is now possible to specify a mode of template or data.  In template
mode, the input is treated as a template and a data file must be
specified if needed.  In data mode, the input is treated as data and a
template file must be specified.  If the input is a directory, it is
walked based on the control file and extension, looking for either
template or data files to apply.